understand revert state and revert_head functions
